+++ dmarc-report/dmarc-report.d/local/share/cook/bin/configure-dmarcreport.sh
@@ -24,6 +24,9 @@ npm i --only=production || true
 
 npm audit fix --force || true
 
+# we'll use this process manager to start the app and background it
+npm install pm2 -g || true
+
 # shellcheck disable=SC3003,SC2039
 # safe(r) separator for sed
 sep=$'\001'
@@ -37,5 +40,11 @@ sep=$'\001'
   sed "s${sep}%%serverport%%${sep}$SERVERPORT${sep}g" \
   > /root/dmarc-report/.env
 
-# start npm
-npm start || true
+# start with process manager
+#npm start || true
+pm2 --name dmarc-report start npm -- start || true
+
+echo "To stop a backgrounded node app, please run:"
+echo ""
+echo "  pm2 delete 0"
+echo ""