+++ dmarc-report/CHANGELOG.md
@@ -1,3 +1,4 @@
 0.0
 
 * First bash at a pot jail with automatic dmarc report by accessing IMAP mailbox
+* Use node pm2 as process manager to background node app